Move state methods to StateMixin
