+++ alibi_detect/utils/state/__init__.py
@@ -1,0 +1,7 @@
+from alibi_detect.utils.state.state import StateMixin, _save_state_dict, _load_state_dict
+
+__all__ = [
+    "StateMixin",
+    "_save_state_dict",
+    "_load_state_dict",
+]
